bug(resolve email in token payload): Removes email in the token's payload basing on lfa feedback
- remove email in the token's payload
- add app.js file in order to make the server file clean
- replace console.log by process.stdout in the server file
- replace status code of 204 by 200
- make changes in the test files in order for them to pass

[finishes #169469835]
