+++ server/tests/user.test.js
@@ -1,6 +1,6 @@
 import chai, { expect } from 'chai';
 import chaiHttp from 'chai-http';
-import app from '../../server';
+import app from '../app';
 
 chai.use(chaiHttp);
